Suppress the final set of warnings in pcl apps

+++ apps/cloud_composer/tools/supervoxels.cpp
@@ -29,7 +29,7 @@ pcl::cloud_composer::SupervoxelsTool::performAction (ConstItemList input_data, P
 {
   if (type != PointTypeFlags::NONE)
   {
-    switch (type)
+    switch ((uint8_t) type)
     {
       case (PointTypeFlags::XYZ | PointTypeFlags::RGB):
         return this->performTemplatedAction<pcl::PointXYZRGB> (input_data);
